% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snakesandladders.R
\name{add_snakesandladders}
\alias{add_snakesandladders}
\title{Applies snake and ladder transitions to the board.}
\usage{
add_snakesandladders(snakematrix, board)
}
\arguments{
\item{snakematrix}{A two column matrix listing the starting points of snakes/ladders in the left column and the endpoints in the right}

\item{board}{A vector of all numbers from 1 to the length of the board.  Index represents board position, value represents what square you end up on in that position}
}
\value{
A new board vector with values replaced corresponding to the locations of snakes and ladders
}
\description{
\code{add_snakesandladders()} applies the start and end points of all of the snakes and ladders on the board.
The input is stored in a two column matrix, which can be assigned manually or through the function \code{snakes_and_ladders_sim}.  
This function changes the numbers on the board to match where each snake or ladders ends.
For example, if a snake leads from square 55 to square 32, then the number 55 in the board vector is replaced by 32.
}
\examples{
add_snakesandladders(snakematrix = default_ladders_and_snakes, board = default_input_board)
}
