% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_probabilities.R
\name{roll_probabilities}
\alias{roll_probabilities}
\title{Runs the simulation for a a range of die sizes}
\usage{
roll_probabilities(maxroll = 100)
}
\arguments{
\item{maxroll}{The maximum dize.  Essentially, the number of iterations of this function.}
}
\value{
A vector containing the max values of the vector returned by \code{finish_game_chance} after each turn
}
\description{
The function was just an additional little experiment I wanted to try, and is not part of the core functions of the package.
\code{roll_probabilities} creates a new transition matrix, runs the markov chain simulation, and produces a new probability distribution
for each die size in the range.  The maximum value of the output of \code{finish_game_chance} is recorded in the vector \code{optimizedroll}.
This produces an interesting graph structure when plotted versus the die size.
}
\examples{
roll_probabilities(maxroll = 100)
}
