% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakes_and_ladders_sim.R
\name{snakes_and_ladders_sim}
\alias{snakes_and_ladders_sim}
\title{Runs the entire Snakes and Ladders Simulation using user specified parameters}
\usage{
snakes_and_ladders_sim()
}
\value{
A plot of the output of \code{finishgamechance} versus the number of turns.  This distribution shows how probable it is to win the game in x number of turns.
}
\description{
This prompts the user to specify all of the necessary parameters, and then runs through \code{add_snakesandladders(), create_transmatrix(), markov_chain(), and finish_game_chance()}.
It then prints the most probable number of turns it will take to win the game, and a graph of the output of \code{finish_game_chance()} versus the number of turns.
This function takes no arguments, and instead prompts the user for all the necessary parameters.
The default settings for the game can also be used.
}
\details{
Try experimenting with different parameter settings to see what you get.
}
\examples{
snakes_and_ladders_sim()
}
